A minor fix in cnn_trainer.py: test_kwargs bug
Fix a bug in aw_nas.final.cnn_trainer during DDP training. 
Line 110, test_kwargs points completely to train_kwargs when the test dataset has no kwargs. 
Line 135, when the sampler adds train_kwargs, the sampler with test_kwargs in the original Line 126 will be overwritten. 
This results in test_queue actually using the sampler of train_queue. 
When the amount of data in the training set is larger than the amount of data in the test set, the index out of bounds will occur when sampling from the test set.